refactor : ui 디렉토리 구조와 절대경로 리팩토링
기존 잘 알 수 없었던 파일 구조를 의미있게 나누었다.
