+++ client/.env
@@ -1,1 +1,2 @@
 SKIP_PREFLIGHT_CHECK=true
+NODE_PATH=src/
