+++ client/src/presentation/components/atomic-reusable/icon-box.tsx
