+++ client/src/presentation/components/atomic-reusable/message-card-image.tsx
