+++ client/src/presentation/components/sung/channel-plus-modal/channel-plus-modal-header.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import LetterXWhite from "../../assets/letter-x-white.png";
-import { IconBox } from "./icon-box";
+import LetterXWhite from "assets/letter-x-white.png";
+import { IconBox } from "presentation/components/atomic-reusable/icon-box";
 
 const ChannelPlusModalHeaderWrapper = styled.section`
   display: flex;
