+++ client/src/presentation/components/sung/message-card/message-card-contents.tsx
