+++ client/src/presentation/pages/snug/sidebar/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { ChannelList } from "./channel-list";
+import { ChannelList } from "presentation/components/sung/channel-list";
 
 const SidebarWrapper = styled.section`
   min-width: 250px;
